Add object details methods
